Update Header, About sections
I decided to change my approach to how I structured the header. I decided to seperate the Header and Hero sections so that the HTML was clearer. I also decided to move the logo out of the nav into its own div. The header is set to position fixed now so that it came over the hero image. To improve nav legibility I set the background to a semi-transparent blue to match the existing brand palette.